authenticate

+++ src/main/kotlin/com/todo/todo/controller/NoteController.kt
@@ -1,46 +1,45 @@
 package com.todo.todo.controller
 
+import com.todo.todo.controller.NoteController.NoteResponse
 import com.todo.todo.database.model.Note
 import com.todo.todo.database.repository.NoteRepository
+import jakarta.validation.Valid
+import jakarta.validation.constraints.NotBlank
 import org.bson.types.ObjectId
-import org.springframework.web.bind.annotation.DeleteMapping
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RequestParam
-import org.springframework.web.bind.annotation.RestController
+import org.springframework.security.core.context.SecurityContextHolder
+import org.springframework.web.bind.annotation.*
 import java.time.Instant
 
 @RestController
 @RequestMapping("/notes")
 class NoteController(
-    private val repository: NoteRepository
+    private val repository: NoteRepository,
+    private val noteRepository: NoteRepository
 ) {
 
     data class NoteRequest(
         val id: String?,
+        @field:NotBlank(message = "Title can't be blank.")
         val title: String,
         val content: String,
         val color: Long,
-        val completed: Boolean,
-        val ownerId: String
+        val completed: Boolean
     )
 
     data class NoteResponse(
-        val id: String?,
+        val id: String,
         val title: String,
         val content: String,
         val color: Long,
-        val completed: Boolean,
-        val createdAt: Instant
+        val createdAt: Instant,
+        val completed: Boolean
     )
 
     @PostMapping
     fun save(
-        @RequestBody body: NoteRequest
+        @Valid @RequestBody body: NoteRequest
     ): NoteResponse {
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         val note = repository.save(
             Note(
                 id = body.id?.let { ObjectId(it) } ?: ObjectId.get(),
@@ -48,7 +47,7 @@ class NoteController(
                 content = body.content,
                 color = body.color,
                 createdAt = Instant.now(),
-                ownerId = ObjectId(),
+                ownerId = ObjectId(ownerId),
                 completed = false
             )
         )
@@ -57,9 +56,8 @@ class NoteController(
     }
 
     @GetMapping
-    fun findByOwnerId(
-        @RequestParam(required = true) ownerId: String
-    ): List<NoteResponse>{
+    fun findByOwnerId(): List<NoteResponse> {
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         return repository.findByOwnerId(ObjectId(ownerId)).map {
             it.toResponse()
         }
@@ -67,19 +65,23 @@ class NoteController(
 
     @DeleteMapping(path = ["/{id}"])
     fun deleteById(@PathVariable id: String) {
-        repository.deleteById(ObjectId(id))
+        val note = noteRepository.findById(ObjectId(id)).orElseThrow {
+            IllegalArgumentException("Note not found")
+        }
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
+        if(note.ownerId.toHexString() == ownerId) {
+            repository.deleteById(ObjectId(id))
+        }
     }
-
-    private fun Note.toResponse(): NoteController.NoteResponse {
-        return NoteResponse(
-            id = id?.toHexString(),
-            title = title,
-            content = content,
-            color = color,
-            createdAt = createdAt,
-            completed = false
-        )
-    }
-
 }
 
+private fun Note.toResponse(): NoteResponse {
+    return NoteResponse(
+        id = id!!.toHexString(),
+        title = title,
+        content = content,
+        color = color,
+        createdAt = createdAt,
+        completed = false
+    )
+}
